Move city tables to a separate component, implement infinity scroll
