+++ src/frontend/main/TableView.tsx
@@ -1,15 +1,11 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
-import handleResponseData from './handleResponseData';
-import { NAVHEIGHT, CITYSELECTHEIGHT } from './constants';
+import CityTables from './CityTables';
+import { CITYSELECTHEIGHT } from './constants';
 
 type TableViewProps = {
   cities: Array<string>;
   data: any;
-  isEmbed: boolean;
-};
-
-type TableViewWrapperProps = {
   isEmbed: boolean;
 };
 
@@ -25,98 +21,6 @@ const CitySelect = styled.div`
   select {
     max-width: 200px;
   }
-`;
-
-const TableViewWrapper = styled.div<TableViewWrapperProps>`
-  max-width: 600px;
-  margin: 0 auto;
-  height: ${props => (props.isEmbed ? `calc(100vh - (${CITYSELECTHEIGHT}px + ${NAVHEIGHT}px))` : 'auto')};
-  overflow: ${props => (props.isEmbed ? 'auto' : 'initial')};
-`;
-
-const TableContainer = styled.div`
-  width: 100%;
-  border-bottom: 1px solid ${props => props.theme.grey};
-  padding: 0 0 24px 0;
-
-  p {
-    margin: 0 0 4px 0;
-  }
-`;
-
-const Table = styled.table`
-  width: 100%;
-  table-layout: fixed;
-  border-collapse: collapse;
-  margin-top: 16px;
-
-  th {
-    font-weight: normal;
-
-    &:nth-child(1) {
-      text-align: left;
-    }
-  }
-
-  th,
-  td {
-    padding: 1px 4px;
-    width: 90px;
-
-    @media (min-width: 450px) {
-      padding-left: 0;
-      width: 140px;
-    }
-  }
-
-  th:nth-child(1),
-  td:nth-child(1) {
-    padding-left: 16px;
-    width: auto;
-  }
-
-  th:nth-child(3),
-  td:nth-child(3) {
-    padding-right: 16px;
-  }
-
-  td {
-    text-align: center;
-  }
-
-  tbody tr:nth-child(1) {
-    th,
-    td {
-      padding-top: 8px;
-    }
-  }
-`;
-
-const TableHead = styled.thead`
-  background: ${props => props.theme.lightGrey};
-
-  th {
-    padding: 6px 4px;
-    font-style: italic;
-    font-size: 14px;
-  }
-`;
-
-const CityHeadingContainer = styled.div`
-  padding: 0 16px;
-`;
-
-const BoldText = styled.p`
-  font-weight: bold;
-`;
-
-const ItalicText = styled.p`
-  font-style: italic;
-`;
-
-const CityHeading = styled.h2`
-  font-size: 21px;
-  margin: 18px 0 8px;
 `;
 
 const Label = styled.label`
@@ -147,61 +51,7 @@ const TableView = ({ data, cities, isEmbed }: TableViewProps) => {
           })}
         </select>
       </CitySelect>
-      <TableViewWrapper isEmbed={isEmbed}>
-        {cityList.map((item: any) => {
-          const formattedData = handleResponseData(item);
-
-          return (
-            <TableContainer key={item.city}>
-              <CityHeadingContainer>
-                <CityHeading>{item.city}</CityHeading>
-                <BoldText>Vastauksia yhteensä {formattedData.responsesTotal}</BoldText>
-
-                {formattedData.percentageOfPopulation != null ? (
-                  <ItalicText>{formattedData.percentageOfPopulation} % väkiluvusta</ItalicText>
-                ) : null}
-              </CityHeadingContainer>
-              <Table>
-                <TableHead>
-                  <tr>
-                    <th>Oireet</th>
-                    <th>Vastauksia</th>
-                    <th>Osuus väkiluvusta</th>
-                  </tr>
-                </TableHead>
-                <tbody>
-                  <tr>
-                    <th scope="row">Epäilys koronavirus&shy;tartunnasta</th>
-                    <td>{formattedData.suspicionTotal}</td>
-                    <td>
-                      {formattedData.suspicionPercentage != null ? `${formattedData.suspicionPercentage} %` : '-'}
-                    </td>
-                  </tr>
-                  <tr>
-                    <th scope="row">Yskää</th>
-                    <td>{formattedData.coughTotal}</td>
-                    <td>{formattedData.coughPercentage != null ? `${formattedData.coughPercentage} %` : '-'}</td>
-                  </tr>
-                  <tr>
-                    <th scope="row">Kuumetta</th>
-                    <td>{formattedData.feverTotal}</td>
-                    <td>{formattedData.feverPercentage != null ? `${formattedData.feverPercentage} %` : '-'}</td>
-                  </tr>
-                  <tr>
-                    <th scope="row">Vaikeuksia hengittää</th>
-                    <td>{formattedData.breathingDifficultiesTotal}</td>
-                    <td>
-                      {formattedData.breathingDifficultiesPercentage != null
-                        ? `${formattedData.breathingDifficultiesPercentage} %`
-                        : '-'}
-                    </td>
-                  </tr>
-                </tbody>
-              </Table>
-            </TableContainer>
-          );
-        })}
-      </TableViewWrapper>
+      <CityTables cityList={cityList} isEmbed={isEmbed} />
     </>
   );
 };
